pyproject.toml
[project]
name = "ad-pipeline"
version = "0.1.0"
description = "Conversational Meta ad pipeline: URL -> copy -> HTML creatives -> Tigris-hosted PNGs -> published Meta ads"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.63",
    "browser-use-sdk>=3.0",
    "weasyprint>=62",
    "pypdfium2>=4",
    "pillow>=11",
    "boto3>=1.35",
    "chainlit>=2.0",
    "pydantic>=2.9",
    "python-dotenv>=1.0",
    "langfuse>=3.0",
]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "moto>=5",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration_live: talks to live external services (Anthropic/Browser Use/Tigris/Pipeboard/Meta sandbox)",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tools"]
